[refactor] make framework adaptor use template pattern instead of adaptor pattern
