+++ example/vulnerable_code/tainted_arg_normal_function.py
@@ -1,0 +1,20 @@
+import os
+
+
+def store_uploaded_file(title, uploaded_file):
+    """ Stores a temporary uploaded file on disk """
+    upload_dir_path = '%s/static/taskManager/uploads' % (
+        os.path.dirname(os.path.realpath(__file__)))
+    if not os.path.exists(upload_dir_path):
+        os.makedirs(upload_dir_path)
+
+    # A1: Injection (shell)
+    os.system(
+        "mv " +
+        uploaded_file.temporary_file_path() +
+        " " +
+        "%s/%s" %
+        (upload_dir_path,
+         title))
+
+    return '/static/taskManager/uploads/%s' % (title)
