+++ pyt/framework_adaptor.py
@@ -1,23 +1,75 @@
-"""A framework adaptor is a adaptor used to adapt
-the source code to a specific framework."""
-from abc import ABCMeta, abstractmethod
+"""A generic framework adaptor that leaves route criteteria to the caller."""
+import ast
 
+from .ast_helper import Arguments
 from .base_cfg import AssignmentNode
+from .interprocedural_cfg import interprocedural
+from .module_definitions import project_definitions
 
 
-class FrameworkAdaptor(metaclass=ABCMeta):
-    """An engine that should be used as base class
-    to specify how to find all sources and sinks."""
+class FrameworkAdaptor():
+    """An engine that should be uses the template pattern
+    to specify how to find all sources and sinks in a framework."""
 
-    def __init__(self, cfg_list, project_modules, local_modules):
+    def __init__(self, cfg_list, project_modules, local_modules, is_route_function):
         self.cfg_list = cfg_list
         self.project_modules = project_modules
         self.local_modules = local_modules
+        self.is_route_function = is_route_function
         self.run()
 
-    @abstractmethod
+    def get_func_cfg_with_tainted_args(self, definition):
+        """Build a function cfg and return it, with all arguments tainted."""
+        func_cfg = interprocedural(definition.node, self.project_modules,
+                                   self.local_modules, definition.path,
+                                   definition.module_definitions)
+
+        args = Arguments(definition.node.args)
+        if args:
+            function_entry_node = func_cfg.nodes[0]
+            function_entry_node.outgoing = []
+            first_node_after_args = func_cfg.nodes[1]
+            first_node_after_args.ingoing = []
+
+            # We're just gonna give all the tainted args the lineno of the def
+            definition_lineno = definition.node.lineno
+
+            # Taint all the arguments
+            for arg in args:
+                tainted_node = TaintedNode(arg, arg,
+                                           None, [],
+                                           line_number=definition_lineno,
+                                           path=definition.path)
+                function_entry_node.connect(tainted_node)
+                # 1 and not 0 for Entry Node to remain first in the list
+                func_cfg.nodes.insert(1, tainted_node)
+
+            first_arg = func_cfg.nodes[len(args)]
+            first_arg.connect(first_node_after_args)
+
+        return func_cfg
+
+    def find_route_functions_taint_args(self):
+        """Find all route functions and taint all of their arguments.
+
+        Yields:
+            CFG of each route function, with args marked as tainted.
+        """
+        for definition in _get_func_nodes():
+            if self.is_route_function(definition.node):
+                yield self.get_func_cfg_with_tainted_args(definition)
+
     def run(self):
-        pass
+        """Run find_route_functions_taint_args on each CFG."""
+        function_cfgs = list()
+        for _ in self.cfg_list:
+            function_cfgs.extend(self.find_route_functions_taint_args())
+        self.cfg_list.extend(function_cfgs)
+
+def _get_func_nodes():
+    """Get all function nodes."""
+    return [definition for definition in project_definitions.values() \
+            if isinstance(definition.node, ast.FunctionDef)]
 
 
 class TaintedNode(AssignmentNode):
