+++ pyt/framework_helper.py
@@ -1,0 +1,24 @@
+"""Provides helper functions that help with determining if a function is a route function."""
+import ast
+from .ast_helper import get_call_names
+
+def is_function(function):
+    """Check whether function uses a route decorator."""
+    return True
+
+
+def is_flask_route_function(ast_node):
+    """Check whether function uses a route decorator."""
+    for decorator in ast_node.decorator_list:
+        if isinstance(decorator, ast.Call):
+            if _get_last_of_iterable(get_call_names(decorator.func)) == 'route':
+                return True
+    return False
+
+
+def _get_last_of_iterable(iterable):
+    """Get last element of iterable."""
+    item = None
+    for item in iterable:
+        pass
+    return item
