+++ tests/vulnerabilities_across_files_test.py
@@ -7,7 +7,8 @@ from pyt.ast_helper import get_call_names_as_string
 from pyt.base_cfg import Node
 from pyt.constraint_table import constraint_table, initialize_constraint_table
 from pyt.fixed_point import analyse
-from pyt.flask_adaptor import FlaskAdaptor
+from pyt.framework_adaptor import FrameworkAdaptor
+from pyt.framework_helper import is_flask_route_function
 from pyt.lattice import Lattice
 from pyt.project_handler import get_directory_modules, get_modules
 from pyt.reaching_definitions_taint import ReachingDefinitionsTaintAnalysis
@@ -24,7 +25,7 @@ class EngineTest(BaseTestCase):
 
         cfg_list = [self.cfg]
 
-        FlaskAdaptor(cfg_list, [], [])
+        FrameworkAdaptor(cfg_list, [], [], is_flask_route_function)
 
         initialize_constraint_table(cfg_list)
